+++ src/main/java/org/folio/services/servicepoint/ServicePointService.java
@@ -1,6 +1,9 @@
 package org.folio.services.servicepoint;
 
 import static io.vertx.core.Future.succeededFuture;
+import static org.folio.rest.impl.ServicePointApi.LOCATION_PREFIX;
+import static org.folio.rest.jaxrs.resource.ServicePoints.PostServicePointsResponse.headersFor201;
+import static org.folio.rest.jaxrs.resource.ServicePoints.PostServicePointsResponse.respond201WithApplicationJson;
 
 import io.vertx.core.Context;
 import io.vertx.core.Future;
@@ -43,6 +46,15 @@ public class ServicePointService {
       .onSuccess(oldServicePoint -> servicePointDomainEventPublisher
         .publishUpdated(oldServicePoint, entity))
       .map(x -> ItemStorage.PutItemStorageItemsByItemIdResponse.respond204());
+  }
+
+  public Future<Response> createServicePoint(String servicePointId, Servicepoint servicePoint) {
+    servicePoint.setId(servicePointId);
+    return servicePointRepository.save(servicePointId, servicePoint)
+      .compose(notUsed ->
+        servicePointDomainEventPublisher.publishCreated(servicePoint)
+          .map(resp -> respond201WithApplicationJson(servicePoint, headersFor201()
+            .withLocation(LOCATION_PREFIX + servicePointId))));
   }
 
   public Future<Boolean> deleteServicePoint(String servicePointId) {